Expose InspectorClient valueSubtype and descriptionForValueSubtype
When the inspector returns a RemoteObject, the proper subType must be returned
in order for the driver (i.e. puppeteer) to know what methods are available.

For example, on any page.$eval, the first  thing puppeteer does is request the
document object via a Runtime.callFunctionOn which simply does

   return document;

If the description of the document that we return (a "RemoteObject") doesn't
include the "node" subtype, the puppeteer will fail.

This PR provides the hooks for an implementation to provide its own valueSubtype
implementation. However, in order for v8 to include the `node` field, we must
always also include a "description" field, which is why this PR also allows
implementations to provide its own `descriptionForValueSubtype`.